requirements.txt
imageio
pilgram
scipy
numpy<1.24>=1.18
torchvision
